[COOK-858] - account for freebsd and archlinux
* Add java_home attribute for both platforms
* Only perform update-alternatives on debian/redhat families

+++ java/recipes/openjdk.rb
@@ -39,43 +39,45 @@ ruby_block  "set-env-java-home" do
   end
 end
 
-ruby_block "update-java-alternatives" do
-  block do
-    if platform?("ubuntu", "debian") and version == 6
-      run_context = Chef::RunContext.new(node, {})
-      r = Chef::Resource::Execute.new("update-java-alternatives", run_context)
-      r.command "update-java-alternatives -s java-6-openjdk"
-      r.returns [0,2]
-      r.run_action(:create)       
-    else
-      # have to do this on ubuntu for version 7 because Ubuntu does
-      # not currently set jdk 7 as the default jvm on installation
-      require "fileutils"
-      arch = node['kernel']['machine'] =~ /x86_64/ ? "x86_64" : "i386"
-      Chef::Log.debug("glob is #{java_home_parent}/java*#{version}*openjdk*")
-      jdk_home = Dir.glob("#{java_home_parent}/java*#{version}*openjdk{,[-\.]#{arch}}")[0]
-      Chef::Log.debug("jdk_home is #{jdk_home}")
-      # delete the symlink if it already exists
-      if File.exists? java_home
-        FileUtils.rm_f java_home
-      end
-      FileUtils.ln_sf jdk_home, java_home
-    
-      cmd = Chef::ShellOut.new(
-         %Q[ update-alternatives --install /usr/bin/java java #{java_home}/bin/java 1;
+if platform?("ubuntu","debian","redhat","centos","fedora","scientific","amazon")
+  ruby_block "update-java-alternatives" do
+    block do
+      if platform?("ubuntu", "debian") and version == 6
+        run_context = Chef::RunContext.new(node, {})
+        r = Chef::Resource::Execute.new("update-java-alternatives", run_context)
+        r.command "update-java-alternatives -s java-6-openjdk"
+        r.returns [0,2]
+        r.run_action(:create)
+      else
+        # have to do this on ubuntu for version 7 because Ubuntu does
+        # not currently set jdk 7 as the default jvm on installation
+        require "fileutils"
+        arch = node['kernel']['machine'] =~ /x86_64/ ? "x86_64" : "i386"
+        Chef::Log.debug("glob is #{java_home_parent}/java*#{version}*openjdk*")
+        jdk_home = Dir.glob("#{java_home_parent}/java*#{version}*openjdk{,[-\.]#{arch}}")[0]
+        Chef::Log.debug("jdk_home is #{jdk_home}")
+        # delete the symlink if it already exists
+        if File.exists? java_home
+          FileUtils.rm_f java_home
+        end
+        FileUtils.ln_sf jdk_home, java_home
+
+        cmd = Chef::ShellOut.new(
+          %Q[ update-alternatives --install /usr/bin/java java #{java_home}/bin/java 1;
              update-alternatives --set java #{java_home}/bin/java  ]
-                             ).run_command
-      unless cmd.exitstatus == 0 or  cmd.exitstatus == 2
-        Chef::Application.fatal!("Failed to update-alternatives for openjdk!")
+          ).run_command
+        unless cmd.exitstatus == 0 or  cmd.exitstatus == 2
+          Chef::Application.fatal!("Failed to update-alternatives for openjdk!")
+        end
       end
     end
+    action :nothing
   end
-  action :nothing
 end
-  
+
 pkgs.each do |pkg|
   package pkg do
     action :install
-    notifies :create, "ruby_block[update-java-alternatives]"
+    notifies :create, "ruby_block[update-java-alternatives]"  if platform?("ubuntu","debian","redhat","centos","fedora","scientific","amazon")
   end
 end
